+++ src/plone.server/plone/server/interfaces/__init__.py
@@ -1,0 +1,141 @@
+# -*- coding: utf-8 -*-
+from plone.server.interfaces.content import IApplication  # noqa
+from plone.server.interfaces.content import IContainer  # noqa
+from plone.server.interfaces.content import IContentNegotiation  # noqa
+from plone.server.interfaces.content import IDataBase  # noqa
+from plone.server.interfaces.content import IItem  # noqa
+from plone.server.interfaces.content import IRegistry  # noqa
+from plone.server.interfaces.content import IResource  # noqa
+from plone.server.interfaces.content import IResourceFactory  # noqa
+from plone.server.interfaces.content import ISite  # noqa
+from plone.server.interfaces.content import IStaticDirectory  # noqa
+from plone.server.interfaces.content import IStaticFile  # noqa
+from plone.server.interfaces.events import INewUserAdded  # noqa
+from plone.server.interfaces.events import IObjectFinallyCreatedEvent  # noqa
+from plone.server.interfaces.files import IFile  # noqa
+from plone.server.interfaces.files import IFileField  # noqa
+from plone.server.interfaces.files import IFileManager  # noqa
+from plone.server.interfaces.files import IStorage  # noqa
+from plone.server.interfaces.files import NotStorable  # noqa
+from plone.server.interfaces.text import IRichText  # noqa
+from plone.server.interfaces.text import IRichTextValue  # noqa
+from plone.server.interfaces.views import ICONNECT  # noqa
+from plone.server.interfaces.views import IDELETE  # noqa
+from plone.server.interfaces.views import IDownloadView  # noqa
+from plone.server.interfaces.views import IGET  # noqa
+from plone.server.interfaces.views import IHEAD  # noqa
+from plone.server.interfaces.views import IOPTIONS  # noqa
+from plone.server.interfaces.views import IPATCH  # noqa
+from plone.server.interfaces.views import IPOST  # noqa
+from plone.server.interfaces.views import IPUT  # noqa
+from plone.server.interfaces.views import ITraversableView  # noqa
+from plone.server.interfaces.views import IView  # noqa
+from zope.i18nmessageid.message import MessageFactory
+from zope.interface import Interface
+
+
+_ = MessageFactory('plone.server')
+
+DEFAULT_ADD_PERMISSION = 'plone.AddContent'
+DEFAULT_READ_PERMISSION = 'plone.ViewContent'
+DEFAULT_WRITE_PERMISSION = 'plone.ManageContent'
+
+CATALOG_KEY = 'plone.server.directives.catalog'
+FIELDSETS_KEY = 'plone.server.directives.fieldsets'
+INDEX_KEY = 'plone.server.directives.index'
+READ_PERMISSIONS_KEY = 'plone.server.directives.read-permissions'
+WRITE_PERMISSIONS_KEY = 'plone.server.directives.write-permissions'
+
+SHARED_CONNECTION = False
+WRITING_VERBS = ['POST', 'PUT', 'PATCH', 'DELETE']
+SUBREQUEST_METHODS = ['get', 'delete', 'head', 'options', 'patch', 'put']
+
+
+class IFormFieldProvider(Interface):
+    """Marker interface for schemata that provide form fields.
+    """
+
+
+class IRequest(Interface):
+    pass
+
+
+class IResponse(Interface):
+
+    def __init__(context, request):
+        pass
+
+
+# Classes as for marker objects to lookup
+
+
+class IRenderFormats(Interface):
+    pass
+
+
+class IFrameFormats(Interface):
+    pass
+
+
+class ILanguage(Interface):
+    pass
+
+
+# Target interfaces on resolving
+
+class IRendered(Interface):
+    pass
+
+
+class ITranslated(Interface):
+    pass
+
+# Get Absolute URL
+
+
+class IAbsoluteURL(Interface):
+    pass
+
+
+# Addon interface
+
+class IAddOn(Interface):
+
+    def install(self, site):
+        pass
+
+    def uninstall(self):
+        pass
+
+
+class TransformError(Exception):
+    """Exception raised if a value could not be transformed. This is normally
+    caused by another exception. Inspect self.cause to find that.
+    """
+
+    def __init__(self, message, cause=None):
+        self.message = message
+        self.cause = cause
+
+    def __str__(self):
+        return self.message
+
+
+class ITransformer(Interface):
+    """A simple abstraction for invoking a transformation from one MIME
+    type to another.
+    This is not intended as a general transformations framework, but rather
+    as a way to abstract away a dependency on the underlying transformation
+    engine.
+    This interface will be implemented by an adapter onto the context where
+    the value is stored.
+    """
+
+    def __init__(object):
+        """Set the value object."""
+
+    def __call__():
+        """Transform the IRichTextValue 'value' to the given MIME type.
+        Return a unicode string. Raises TransformError if something went
+        wrong.
+        """
