+++ src/plone.server/plone/server/interfaces/views.py
@@ -1,0 +1,52 @@
+from zope.interface import Interface
+
+
+class IView(Interface):
+
+    def __init__(context, request):
+        pass
+
+    async def __call__(self):
+        pass
+
+
+class ITraversableView(IView):
+
+    def publishTraverse(traverse_to):
+        pass
+
+
+class IDownloadView(IView):
+    pass
+
+
+class IGET(IView):
+    pass
+
+
+class IPUT(IView):
+    pass
+
+
+class IPOST(IView):
+    pass
+
+
+class IPATCH(IView):
+    pass
+
+
+class IDELETE(IView):
+    pass
+
+
+class IOPTIONS(IView):
+    pass
+
+
+class IHEAD(IView):
+    pass
+
+
+class ICONNECT(IView):
+    pass
